[chore] 프로젝트 그룹 구조 변경 및 코드 리팩토링
