+++ Project17-C-Map/InteractiveClusteringMap/Extension/NMFMarker+Animation.swift
@@ -1,13 +1,11 @@
 //
-//  NMFMarker+.swift
+//  NMFMarker+Animation.swift
 //  InteractiveClusteringMap
 //
-//  Created by A on 2020/12/10.
+//  Created by Seungeon Kim on 2020/12/14.
 //
 
-import Foundation
 import NMapsMap
-
 
 extension NMFMarker {
     
@@ -16,3 +14,4 @@ extension NMFMarker {
     }
     
 }
+
