+++ Project17-C-Map/InteractiveClusteringMap/UI/View/Map/LeafNodeMarkerInfoWindowView.xib
@@ -8,14 +8,14 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CustomInfoWindow">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="LeafNodeMarkerInfoWindowView" customModule="InteractiveClusteringMap" customModuleProvider="target">
             <connections>
                 <outlet property="categoryLabel" destination="pGH-yI-kfA" id="wn3-qB-3nQ"/>
                 <outlet property="titleLabel" destination="GMb-lm-Ykj" id="ar4-gb-EFp"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" restorationIdentifier="CustomInfoWindow" id="iN0-l3-epB">
+        <view contentMode="scaleToFill" restorationIdentifier="LeafNodeMarkerInfoWindow" id="iN0-l3-epB">
             <rect key="frame" x="0.0" y="0.0" width="373" height="105"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
