+++ Project17-C-Map/InteractiveClusteringMap/UI/View/PlaceInfo/PlaceCell.swift
